// Models/assign-transport-model.ts
export class AssignTransportModel {
    BranchId:number = 0;
    CompId: number = 0;
    AssignTransMId :number = 0;
    InvoiceId: string = "";
    TransportModeId: string = "";
    PersonName: string = "";
    PersonAddress: string = "";
    PersonMobileNo: string = "";
    OtherDetails: string = "";
    TransporterId: string = "";
    TransporterName:String="";
    CourierName:string="";
    Delivery_Remark: string = "";
    CAId: string = "";
    CourierId: string = "";
    isActive: string = "";
    Action: string = "";
    Addedby: string = "";
    AttachedInvId :number =0;
}